+++ src/routes/auth/discord/callback.js
@@ -30,7 +30,7 @@ const DiscordAuth = () => {
 
     useEffect(() => {
         let callback_url = `https://${vars.host}/auth/discord/callback`;
-        if (vars.discordClientID === 1120997206938361877) {
+        if (appContext.apiConfig.discord_client_id === 1120997206938361877) {
             callback_url = `https://oauth.chub.page/discord-auth`;
         }
         async function validateDiscordAuth() {