+++ src/routes/auth/login.js
@@ -1,7 +1,7 @@
 import { useState, useCallback, useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import { ThemeContext } from '../../context';
+import { AppContext, ThemeContext } from '../../context';
 
 import { Button, Card, Grid, Typography, TextField, CardContent, ButtonGroup, Box, IconButton, Dialog, DialogTitle, DialogContent, Snackbar, Alert, useMediaQuery } from '@mui/material';
 import { Portal } from '@mui/base';
@@ -18,6 +18,7 @@ var vars = require("../../variables");
 
 const AuthLogin = () => {
     const { t: tr } = useTranslation();
+    const { apiConfig } = useContext(AppContext);
     const { themeSettings } = useContext(ThemeContext);
 
     const CONNECTION_NAME = { "email": tr("email"), "discord": "Discord", "steam": "Steam", "truckersmp": "TruckersMP" };
@@ -116,7 +117,7 @@ const AuthLogin = () => {
                             <Typography variant="body2" onClick={() => { if (validateEP()) { setAction("reset-password"); setModalCaptcha(true); } }} sx={{ cursor: "pointer", width: "fit-content" }}>{tr("forgot_your_password")}</Typography>
                             <br />
                             <ButtonGroup fullWidth>
-                                <Button variant="contained" color="primary" disabled={authDisabled || !vars.apiconfig.register_methods.includes("email")} onClick={() => { if (validateEP()) { setAction("register"); setModalCaptcha(true); } }}>{tr("register")}</Button>
+                                <Button variant="contained" color="primary" disabled={authDisabled || !apiConfig.register_methods.includes("email")} onClick={() => { if (validateEP()) { setAction("register"); setModalCaptcha(true); } }}>{tr("register")}</Button>
                                 <Button variant="contained" color="info" disabled={authDisabled} onClick={() => { if (validateEP()) { setAction("login"); setModalCaptcha(true); } }}>{tr("login")}</Button>
                             </ButtonGroup>
                         </Grid>
@@ -126,7 +127,7 @@ const AuthLogin = () => {
                                 <img src={`https://cdn.chub.page/assets/${vars.dhconfig.abbr}/banner.png?${vars.dhconfig.banner_key !== undefined ? vars.dhconfig.banner_key : ""}`} alt="" style={{ width: "100%" }} />
                             </Box>
                             <br />
-                            <Typography variant="body2" sx={{ mb: "5px" }}>{tr("register_with")}: {vars.apiconfig.register_methods.map((item) => (CONNECTION_NAME[item])).join(" / ")}</Typography>
+                            <Typography variant="body2" sx={{ mb: "5px" }}>{tr("register_with")}: {apiConfig.register_methods.map((item) => (CONNECTION_NAME[item])).join(" / ")}</Typography>
                             <ButtonGroup fullWidth>
                                 <Button variant="contained" color="primary" onClick={() => { navigate("/auth/discord/redirect"); }}>
                                     <FontAwesomeIcon icon={faDiscord} />&nbsp;&nbsp;Discord