+++ changelog.md
@@ -40,6 +40,7 @@
     - Improved handling of theme update
 11. Replaced `vars.dlogDetails` with Context API `dlogDetailsCache`, also supported background loading
 12. Replaced `vars.userBanner` with Context API `curUserBanner`
+13. Replaced `vars.apiconfig` with Context API `apiConfig`
 
 ## v3.3.x